refactor(common/lmlayer): move default word-breaker to @keymanapp/lexical-model-word-breakers
