+++ common/lexical-model-word-breakers/src/index.ts
@@ -1,2 +1,5 @@
 import {ascii} from "./ascii-word-breaker";
+import {default_} from "./default";
+
 export {ascii};
+export {default_ as default};
